requirements.txt
aiogram==3.1.1
APScheduler==3.10.4
python-dotenv==1.0.0
aiosqlite==0.20.0
geopy==2.4.1
timezonefinder==6.5.7